# api_docs.yaml
openapi: 3.0.0
info:
  title: LaunchPad Placement Portal API
  version: 1.0.0
  description: |
    REST API for the LaunchPad campus placement portal.

    **Authentication:** Most endpoints require an `Authentication-Token` header.
    Log in via `/auth/login` to get your token, then click "Authorize" above and paste it.
servers:
  - url: http://localhost:5001/api
paths:
...
  /company/profile:
    get:
      summary: Get company profile
      tags: [Company]
      security:
        - TokenAuth: []
      responses:
        "200": { description: Company profile data }
    put:
      summary: Update company profile
      tags: [Company]
      security:
        - TokenAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCompanyProfile"
            example:
              name: Google LLC
              description: Updated company description
              website: https://google.com
              location: Mountain View, CA
              industry: Technology
      responses:
        "200": { description: Profile updated }
  /company/drives:
    get:
      summary: List company's drives
      tags: [Company]
      security:
        - TokenAuth: []
      responses:
        "200": { description: List of drives }
    post:
      summary: Create a placement drive
      description: Company must be approved. Drive is created in pending state awaiting admin approval.
      tags: [Company]
      security:
        - TokenAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDrive"
            example:
              title: Backend Engineer 2026
              role_offered: Software Engineer
              package_lpa: 15.0
              location: Bangalore
              eligibility_cgpa: 7.5
              eligible_branches: ["Computer Science", "Electronics"]
              deadline: "2026-12-31T23:59:59+05:30"
              description: Looking for strong backend engineers with Python/Go experience.
      responses:
        "201": { description: Drive created (pending approval) }
        "400": { description: Validation error }
        "403": { description: Company not approved or blacklisted }
  /company/drives/{drive_id}:
    put:
      summary: Update a drive
      tags: [Company]
      security:
        - TokenAuth: []
      parameters:
        - { name: drive_id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateDrive"
            example:
              description: Updated drive description
              status: closed
      responses:
        "200": { description: Drive updated }
        "403": { description: Not your drive }
  /company/drives/{drive_id}/applications:
    get:
      summary: List applications for a drive
      tags: [Company]
      security:
        - TokenAuth: []
      parameters:
        - { name: drive_id, in: path, required: true, schema: { type: integer } }
      responses:
        "200": { description: Applications list }
        "403": { description: Access denied }
  /company/applications/{app_id}/status:
    put:
      summary: Update application status
      description: |
        Valid transitions:
        - applied → shortlisted, rejected
        - shortlisted → selected, rejected
        When setting status to `selected`, `joining_date` is required.
        When setting status to `rejected`, `feedback` is required.
      tags: [Company]
      security:
        - TokenAuth: []
      parameters:
        - { name: app_id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status:
                  type: string
                  enum: [shortlisted, rejected, selected]
                joining_date:
                  type: string
                  format: date
                  description: Required when status is `selected` (YYYY-MM-DD)
                feedback:
                  type: string
                  description: Required when status is `rejected`
            example:
              status: shortlisted
      responses:
        "200": { description: Status updated }
        "400": { description: Invalid transition }
  /company/applications/{app_id}/interview:
    post:
      summary: Schedule interview
      description: Only for shortlisted applications. One interview per application.
      tags: [Company]
      security:
        - TokenAuth: []
      parameters:
        - { name: app_id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScheduleInterview"
            example:
              scheduled_at: "2026-06-15T10:00:00+05:30"
              location: Google Meet
              interview_type: online
              notes: Technical screening round
      responses:
        "201": { description: Interview scheduled }
        "400": { description: Application not shortlisted }
        "409": { description: Interview already scheduled }
  /company/applications/export:
    post:
      summary: Export company applications and placements as CSV via email
      description: Triggers a Celery task. CSV is sent to the company's email.
      tags: [Company]
      security:
        - TokenAuth: []
      responses:
        "200": { description: Export task queued }

  # Student
...
components:
  securitySchemes:
    TokenAuth:
      type: apiKey
      in: header
      name: Authentication-Token
      description: Paste the token returned by POST /auth/login
  schemas:
...
    UpdateCompanyProfile:
      type: object
      properties:
        name: { type: string }
        description: { type: string }
        website: { type: string }
        location: { type: string }
        industry: { type: string }
...
    CreateDrive:
      type: object
      required: [title, deadline]
      properties:
        title: { type: string }
        description: { type: string }
        role_offered: { type: string }
        package_lpa: { type: number }
        location: { type: string }
        eligibility_cgpa: { type: number, default: 0 }
        eligible_branches:
          type: array
          items: { type: string }
        deadline: { type: string, format: date-time, description: "ISO 8601 with timezone" }
    UpdateDrive:
      type: object
      properties:
        title: { type: string }
        description: { type: string }
        role_offered: { type: string }
        package_lpa: { type: number }
        location: { type: string }
        eligibility_cgpa: { type: number }
        eligible_branches:
          type: array
          items: { type: string }
        deadline: { type: string, format: date-time }
        status: { type: string, enum: [open, closed, cancelled] }
    ScheduleInterview:
      type: object
      required: [scheduled_at]
      properties:
        scheduled_at: { type: string, format: date-time, description: "ISO 8601 with timezone" }
        location: { type: string }
        interview_type: { type: string, enum: [in-person, online], default: in-person }
        notes: { type: string }